chore : Splited the database file to multiple files to reduce length and make it easier to maintain
