+++ lib/services/db/upgrade/db_upgrade.dart
@@ -1,0 +1,5 @@
+import 'package:sqflite/sqlite_api.dart';
+
+abstract class DbUpgrade {
+  Future<void> upgrade(Database db, int oldVersion, int newVersion);
+}
